Little refactor, add tests
