+++ console/main.cpp
@@ -1,8 +1,9 @@
 #include <iostream>
 
 #include <Fst.h>
-#include <types.h>
+#include <surfTypes.h>
 
+#include <test.h>
 #include <Arguments.h>
 bool processArgs(Args &args, int argc, char *argv[]) noexcept {
     try {
@@ -23,13 +24,61 @@ bool processArgs(Args &args, int argc, char *argv[]) noexcept {
     return true;
 }
 
-void checkKey(Surf::Fst &fst, const std::string& key) {
-    Surf::value_t* value = fst.exactKeySearch(key);
-    if(value) {
+bool checkKey(Surf::Fst &fst, const std::string &key) {
+    Surf::value_t *value = fst.exactKeySearch(key);
+    if (value) {
         std::cout << "Find [" << key << "], value: " << *value << '\n';
+        return true;
     } else {
         std::cout << "No   [" << key << "]\n";
+        return false;
     }
+}
+
+void runTest() {
+    using Surf::key_t;
+    using Surf::value_t;
+
+    std::vector<std::pair<key_t, value_t>> input{
+            {"hello", 1}, {"he", 3}, {"andatra", 77}, {"a", 15}};
+
+    Surf::Fst fst(input);
+
+    std::vector<key_t> check_keys{"a",
+                                  "many",
+                                  "coco",
+                                  "hello",
+                                  "andatra",
+                                  "yerfbierb",
+                                  "123124234",
+                                  "gefrgrg",
+                                  "a",
+                                  "a",
+                                  "a$",
+                                  "andatra"};
+
+    test(fst, check_keys, input);
+}
+
+void defaultRun() {
+    using Surf::key_t;
+    using Surf::value_t;
+
+    std::vector<std::pair<key_t, value_t>> input{
+            {"hello", 1}, {"he", 3}, {"andatra", 77}, {"a", 15}};
+
+    Surf::Fst fst(input);
+
+    checkKey(fst, "a");
+    checkKey(fst, "an");
+    checkKey(fst, "anda");
+    checkKey(fst, "az");
+    checkKey(fst, "hello");
+    checkKey(fst, "h");
+    checkKey(fst, "he");
+    checkKey(fst, "ell");
+    checkKey(fst, "andatra");
+    checkKey(fst, "rrrere");
 }
 
 int main(int argc, char *argv[]) {
@@ -46,25 +95,12 @@ int main(int argc, char *argv[]) {
         return 0;
     }
 
-    std::vector<std::pair<std::string, u64>> input{
-        {"hello", 1},
-        {"he", 3},
-        {"andatra", 77},
-        {"a",15}
-    };
+    if (args.has(ARG_TEST)) {
+        runTest();
+        return 0;
+    }
 
-    Surf::Fst fst(input);
-
-    checkKey(fst, "a");
-    checkKey(fst, "an");
-    checkKey(fst, "anda");
-    checkKey(fst, "az");
-    checkKey(fst, "hello");
-    checkKey(fst, "h");
-    checkKey(fst, "he");
-    checkKey(fst, "ell");
-    checkKey(fst, "andatra");
-    checkKey(fst, "rrrere");
+    defaultRun();
     std::cout << "Work done, shutdown application" << std::endl;
     return 0;
 }
