+++ console/src/Arguments.h
@@ -1,9 +1,12 @@
 #include <Application/Args.h>
 
-#define ARG_LONG_HELP "help"
-#define ARG_HELP "h"
+#define ARG_HELP "help"
+#define ARG_HELP_SHORT "h"
 
+#define ARG_TEST "test"
+#define ARG_TEST_SHORT "t"
 
 void buildArgs(Args &args) {
-    args.add({ARG_LONG_HELP, ARG_HELP}, "Print help message");
+    args.add({ARG_HELP, ARG_HELP_SHORT}, "Print help message");
+    args.add({ARG_TEST, ARG_TEST_SHORT}, "Run assert tests");
 }
