+++ core/includes/Fst.h
@@ -2,52 +2,34 @@
 
 #include <vector>
 
-#include <Louds/Dense.h>
 #include <surfTypes.h>
-
-#include <iostream>
 
 namespace Surf {
 
+// pre-definition
+inline namespace PrefixTree {
+struct Node;
+}
+
+inline namespace Louds {
+class Dense;
+class Sparse;
+} // namespace Louds
+
 class Fst {
 public:
-    Fst(const std::vector<std::pair<key_t, value_t>> &data) {
-        PrefixTree::Tree buf_tree;
-        for (auto &&pair : data) {
-            buf_tree.insert(pair.first, pair.second);
-        }
+    value_t *exactKeySearch(const key_t &key);
 
-        for (u64 lvl = 0; lvl < buf_tree.size(); ++lvl) {
-            // std::cout << "Insert to lvl " << lvl << '\n';
-            insertToDense(buf_tree.m_head, lvl);
-        }
-    }
-
-    value_t *exactKeySearch(const key_t &key) {
-        return m_dense.exactKeySearch(key.begin(), --key.end());
-    }
+    Fst(const std::vector<std::pair<key_t, value_t>> &data);
+    ~Fst();
 
 private:
-    void insertToDense(Tree::Node *node, u64 lvl, u64 cur_lvl = 0) {
-        if (cur_lvl > lvl) return;
-        if (cur_lvl < lvl) {
-            if (!node->hasChild()) {
-                return;
-            }
-            for (auto &child : *node->data) {
-                if (child) insertToDense(child, lvl, cur_lvl + 1);
-            }
-            return;
-        }
+    void insertToDense(PrefixTree::Node *node, u64 lvl, u64 cur_lvl = 0);
 
-        if (node->hasChild()) {
-            m_dense.addNode(node);
-        }
-    }
-
+    // data
 private:
-    Louds::Dense m_dense;
-    // Louds::Sparce;
+    Louds::Dense *m_dense;
+    // Louds::Sparse *m_sparse;
 };
 
 } // namespace Surf
